[fix] Remove refetch property for album photosboard + move back selector to reducers index file
